coherent tags names

+++ src/vlCore/MemoryFile.cpp
@@ -101,10 +101,9 @@ bool MemoryFile::seekSet_Implementation(long long offset)
 {
   if (!isOpen())
     return false;
-  mPtr = offset;
+  mPtr = (long long)offset;
   if (mPtr < 0) mPtr = 0;
-  if (mPtr > static_cast<long long>(mBuffer->bytesUsed()))
-	  mPtr = mBuffer->bytesUsed();
+  if (mPtr > mBuffer->bytesUsed()) mPtr = mBuffer->bytesUsed();
   return true;
 }
 //-----------------------------------------------------------------------------
